Moved png image data stuff to lib file
